+++ src/config.ts
@@ -12,6 +12,14 @@ import {
   OracleRollMessage,
 } from './module/rolls'
 import { AssetCompendiumBrowser } from './module/item/asset-compendium-browser'
+import Mitt, { Emitter } from 'mitt'
+
+export type EmitterEvents = {
+  closeApp: void
+  highlightMove: string // Foundry ID
+  highlightOracle: string // DF ID
+}
+export type IronswornEmitter = Emitter<EmitterEvents>
 
 export interface IronswornConfig {
   actorClass: typeof IronswornActor
