+++ src/module/vue/provisions.ts
@@ -2,8 +2,7 @@ import { InjectionKey } from 'vue'
 import { enrichHtml, enrichMarkdown } from './vue-plugin'
 import { IronswornActor } from '../actor/actor'
 import { IronswornItem } from '../item/item'
-import mitt, { Emitter } from 'mitt'
-import { Move } from '../features/custommoves'
+import { IronswornEmitter } from '../../config'
 
 // Provided by the Vue plugin
 export const $EnrichHtmlKey = Symbol('$enrichHtml') as InjectionKey<
@@ -13,13 +12,6 @@ export const $EnrichMarkdownKey = Symbol('$enrichMarkdown') as InjectionKey<
   typeof enrichMarkdown
 >
 
-// Typed events
-export type EmitterEvents = {
-  closeApp: void
-  highlightMove: string // Foundry ID
-  highlightOracle: string // DF ID
-}
-export type IronswornEmitter = Emitter<EmitterEvents>
 export const $EmitterKey = Symbol('$emitter') as InjectionKey<IronswornEmitter>
 
 // Sheets have to provide these
